+++ CHANGELOG.md
@@ -248,3 +248,4 @@
 
 # 0.3.0
 * Rename `Root` datatype to `Location` with `Head` (stack) and `Root` (tree) constructors
+* Define `represent` as lens in `Representable` containers